# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IndoLegalBench API",
    "description": "API IndoLegalBench, platform penulisan dan pengukuran test case hukum untuk Veritask. Dokumentasi ini dihasilkan otomatis dari schema Pydantic dan menjadi kontrak resmi antara server dan client.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/suites/{suite_id}/cases": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Buat kasus hukum dalam status draft",
        "description": "Create a draft case. Authors and admins only.",
        "operationId": "create_case_suites__suite_id__cases_post",
        "parameters": [
          {
            "name": "suite_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Suite Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseWrite"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseRead"
                }
              }
            }
          },
          "403": {
            "description": "Bukan pembuat kasus dan bukan admin.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Kasus atau suite tidak ditemukan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Kode kasus sudah dipakai (`CASE_CODE_TAKEN`), kode diubah setelah ada versi yang disetujui (`CASE_CODE_LOCKED`), versi yang disetujui atau sedang ditinjau diubah (`VERSION_LOCKED`), versi baru diminta saat draf atau tinjauan masih berjalan (`VERSION_IN_PROGRESS`), atau versi baru diminta sebelum ada versi yang disetujui (`NO_APPROVED_VERSION`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Suite tidak aktif (`SUITE_NOT_ACTIVE`), tag kosong (`SPLIT_TAG_REQUIRED`), pola case_code ditolak (`CASE_CODE_INVALID`), atau field wajib kosong (`field`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "Daftar ringkas kasus di dalam satu suite",
        "description": "List the short case summary for one suite.",
        "operationId": "list_cases_suites__suite_id__cases_get",
        "parameters": [
          {
            "name": "suite_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Suite Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CaseStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "split_tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SplitTag"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Split Tag"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CaseSummary"
                  },
                  "title": "Response List Cases Suites  Suite Id  Cases Get"
                }
              }
            }
          },
          "404": {
            "description": "Kasus atau suite tidak ditemukan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cases/{case_id}": {
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "Detail satu kasus",
        "description": "Return the full case.",
        "operationId": "get_case_cases__case_id__get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseRead"
                }
              }
            }
          },
          "404": {
            "description": "Kasus atau suite tidak ditemukan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "cases"
        ],
        "summary": "Ubah kasus",
        "description": "Update the open version. The service rejects an approved version.",
        "operationId": "update_case_cases__case_id__put",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseWrite"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseRead"
                }
              }
            }
          },
          "403": {
            "description": "Bukan pembuat kasus dan bukan admin.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Kasus atau suite tidak ditemukan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Kode kasus sudah dipakai (`CASE_CODE_TAKEN`), kode diubah setelah ada versi yang disetujui (`CASE_CODE_LOCKED`), versi yang disetujui atau sedang ditinjau diubah (`VERSION_LOCKED`), versi baru diminta saat draf atau tinjauan masih berjalan (`VERSION_IN_PROGRESS`), atau versi baru diminta sebelum ada versi yang disetujui (`NO_APPROVED_VERSION`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Suite tidak aktif (`SUITE_NOT_ACTIVE`), tag kosong (`SPLIT_TAG_REQUIRED`), pola case_code ditolak (`CASE_CODE_INVALID`), atau field wajib kosong (`field`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/cases/{case_id}/versions": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Buat versi draf dari kasus yang sudah disetujui",
        "description": "Copy the approved wording into a new draft. No request body.",
        "operationId": "start_case_version_cases__case_id__versions_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseRead"
                }
              }
            }
          },
          "403": {
            "description": "Bukan pembuat kasus dan bukan admin.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Kasus atau suite tidak ditemukan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Kode kasus sudah dipakai (`CASE_CODE_TAKEN`), kode diubah setelah ada versi yang disetujui (`CASE_CODE_LOCKED`), versi yang disetujui atau sedang ditinjau diubah (`VERSION_LOCKED`), versi baru diminta saat draf atau tinjauan masih berjalan (`VERSION_IN_PROGRESS`), atau versi baru diminta sebelum ada versi yang disetujui (`NO_APPROVED_VERSION`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Suite tidak aktif (`SUITE_NOT_ACTIVE`), tag kosong (`SPLIT_TAG_REQUIRED`), pola case_code ditolak (`CASE_CODE_INVALID`), atau field wajib kosong (`field`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "Riwayat versi satu kasus",
        "description": "List every stored version, oldest number first.\n\nAuthor, Reviewer, Admin, and Viewer. SCRUM-138 shows this history to\nViewer, and a Reviewer needs it while a new version is in review.",
        "operationId": "list_case_versions_cases__case_id__versions_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VersionSummary"
                  },
                  "title": "Response List Case Versions Cases  Case Id  Versions Get"
                }
              }
            }
          },
          "404": {
            "description": "Kasus tidak ditemukan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cases/{case_id}/versions/compare": {
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "Bandingkan dua nomor versi",
        "description": "Diff two version numbers of one case. The diff is computed on request.\n\nAuthor, Reviewer, Admin, and Viewer. SCRUM-138 shows this compare to\nViewer, and a Reviewer needs it while a new version is in review.",
        "operationId": "compare_case_versions_cases__case_id__versions_compare_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          },
          {
            "name": "a",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Nomor versi pertama",
              "title": "A"
            },
            "description": "Nomor versi pertama"
          },
          {
            "name": "b",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Nomor versi kedua",
              "title": "B"
            },
            "description": "Nomor versi kedua"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionCompare"
                }
              }
            }
          },
          "404": {
            "description": "Kasus atau nomor versi tidak ditemukan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cases/{case_id}/completeness": {
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "Indikator kelengkapan satu kasus",
        "description": "Return what the case still needs before it can be sent for review.",
        "operationId": "get_completeness_cases__case_id__completeness_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseCompleteness"
                }
              }
            }
          },
          "404": {
            "description": "Kasus atau suite tidak ditemukan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActorRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "ActorRead",
        "description": "Who wrote a version, or who froze a suite snapshot."
      },
...
      "AnswerCriteria": {
        "properties": {
          "must_contain": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Must Contain"
          },
          "must_not_contain": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Must Not Contain"
          },
          "expected_conclusion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Conclusion"
          }
        },
        "type": "object",
        "title": "AnswerCriteria",
        "description": "Phrases the answer must or must not contain, plus an expected conclusion."
      },
...
      "CaseCompleteness": {
        "properties": {
          "is_complete": {
            "type": "boolean",
            "title": "Is Complete"
          },
          "ready_for_review": {
            "type": "boolean",
            "title": "Ready For Review"
          },
          "pct": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Pct"
          },
          "missing": {
            "items": {
              "$ref": "#/components/schemas/CompletenessIssue"
            },
            "type": "array",
            "title": "Missing"
          },
          "trap_count": {
            "type": "integer",
            "title": "Trap Count"
          },
          "legal_ref_count": {
            "type": "integer",
            "title": "Legal Ref Count"
          }
        },
        "type": "object",
        "required": [
          "is_complete",
          "ready_for_review",
          "pct",
          "missing",
          "trap_count",
          "legal_ref_count"
        ],
        "title": "CaseCompleteness",
        "description": "Indikator kelengkapan satu kasus (SCRUM-107)."
      },
      "CaseIdentity": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "description": "Judul pertanyaan"
          },
          "question": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 1,
            "title": "Question",
            "description": "Pertanyaan hukum"
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          }
        },
        "type": "object",
        "required": [
          "title",
          "question"
        ],
        "title": "CaseIdentity",
        "description": "Title, question, and optional category for one case."
      },
      "CaseRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "suite_id": {
            "type": "string",
            "format": "uuid",
            "title": "Suite Id"
          },
          "case_code": {
            "type": "string",
            "title": "Case Code"
          },
          "identity": {
            "$ref": "#/components/schemas/CaseIdentity"
          },
          "legal_refs": {
            "items": {
              "$ref": "#/components/schemas/LegalRef"
            },
            "type": "array",
            "title": "Legal Refs"
          },
          "answer_criteria": {
            "$ref": "#/components/schemas/AnswerCriteria"
          },
          "traps": {
            "items": {
              "$ref": "#/components/schemas/Trap"
            },
            "type": "array",
            "title": "Traps"
          },
          "split_tag": {
            "$ref": "#/components/schemas/SplitTag"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "completeness_pct": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Completeness Pct"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "suite_id",
          "case_code",
          "identity",
          "legal_refs",
          "answer_criteria",
          "traps",
          "split_tag",
          "status",
          "completeness_pct",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "CaseRead",
        "description": "Full case returned by create, read, and update."
      },
      "CaseStatus": {
        "type": "string",
        "enum": [
          "draft",
          "in_review",
          "needs_revision",
          "approved"
        ],
        "title": "CaseStatus",
        "description": "Server-owned lifecycle. Clients do not send this on create or update."
      },
      "CaseSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "case_code": {
            "type": "string",
            "title": "Case Code"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "split_tag": {
            "$ref": "#/components/schemas/SplitTag"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "completeness_pct": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Completeness Pct"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "case_code",
          "title",
          "split_tag",
          "status",
          "completeness_pct",
          "updated_at"
        ],
        "title": "CaseSummary",
        "description": "Short list item. The list endpoint does not return the full body."
      },
      "CaseWrite": {
        "properties": {
          "case_code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 2,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{1,63}$",
            "title": "Case Code",
            "description": "Pola sementara, bukan pola final: diawali huruf atau angka, lalu huruf, angka, titik, garis bawah, atau tanda hubung."
          },
          "identity": {
            "$ref": "#/components/schemas/CaseIdentity"
          },
          "legal_refs": {
            "items": {
              "$ref": "#/components/schemas/LegalRef"
            },
            "type": "array",
            "minItems": 1,
            "title": "Legal Refs",
            "description": "Minimal satu rujukan, masing-masing sampai level pasal."
          },
          "answer_criteria": {
            "$ref": "#/components/schemas/AnswerCriteria"
          },
          "traps": {
            "items": {
              "$ref": "#/components/schemas/Trap"
            },
            "type": "array",
            "title": "Traps"
          },
          "split_tag": {
            "$ref": "#/components/schemas/SplitTag"
          }
        },
        "type": "object",
        "required": [
          "case_code",
          "identity",
          "legal_refs",
          "split_tag"
        ],
        "title": "CaseWrite",
        "description": "Body for create and update. The server sets status; clients cannot send it."
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "examples": [
              "No platform account is mapped to this identity."
            ]
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field",
            "description": "Lokasi field yang gagal, misalnya legal_refs[0].pasal"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SplitTag": {
        "type": "string",
        "enum": [
          "dev",
          "test"
        ],
        "title": "SplitTag",
        "description": "Dataset split stored on a version. Only dev and test are valid."
      },
...
      "VersionCompare": {
        "properties": {
          "a": {
            "$ref": "#/components/schemas/VersionSide"
          },
          "b": {
            "$ref": "#/components/schemas/VersionSide"
          },
          "changed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Changed"
          }
        },
        "type": "object",
        "required": [
          "a",
          "b",
          "changed"
        ],
        "title": "VersionCompare",
        "description": "GET /cases/{id}/versions/compare. a and b are version numbers."
      },
...
      "VersionSummary": {
        "properties": {
          "version_no": {
            "type": "integer",
            "title": "Version No"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "author": {
            "$ref": "#/components/schemas/ActorRead"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "changed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Changed"
          }
        },
        "type": "object",
        "required": [
          "version_no",
          "status",
          "author",
          "created_at",
          "changed"
        ],
        "title": "VersionSummary",
        "description": "One row of GET /cases/{id}/versions."
      }
    }
  }
}